feat(fragment): write account sidebar in tsx
