+++ packages/manager/apps/hub/src/index.html
@@ -38,8 +38,9 @@
             data-sidebar-links="{}"
         >
         </ovh-manager-navbar>
-        <ovh-manager-account-sidebar expand="true" me="$ctrl.user">
-        </ovh-manager-account-sidebar>
+        <!-- <ovh-manager-account-sidebar expand="true" me="$ctrl.user">
+        </ovh-manager-account-sidebar> -->
+        <ovh-fragment fragment-id="account-sidebar"></ovh-fragment>
         <ovh-manager-notifications-sidebar> </ovh-manager-notifications-sidebar>
         <div class="position-relative w-100 h-100 overflow-auto ng-cloak">
             <div
