+++ packages/manager/fragments/account-sidebar/src/App.vue
@@ -4,7 +4,7 @@
 
 <script lang="ts">
 import { defineComponent } from 'vue';
-import AccountSidebar from './components/AccountSidebar.vue';
+import AccountSidebar from './components/AccountSidebar';
 
 export default defineComponent({
   name: 'App',
