+++ packages/manager/fragments/account-sidebar/src/styles/account-sidebar-panel.scss
@@ -1,0 +1,38 @@
+.manager-hub-user-panel {
+  @import '~@ovh-ux/ui-kit/dist/scss/_tokens.scss';
+  @import '~bootstrap4/scss/_functions.scss';
+  @import '~bootstrap4/scss/_variables.scss';
+  @import '~bootstrap4/scss/_mixins.scss';
+  @import '~bootstrap4/scss/_utilities.scss';
+  @import '~bootstrap4/scss/_grid.scss';
+  @import '~bootstrap4/scss/_buttons.scss';
+  @import '~bootstrap4/scss/_button-group.scss';
+  @import '~@ovh-ux/manager-hub/src/variables.scss';
+
+  width: 18.75rem;
+  padding-left: 2.75rem;
+  background-color: $p-075;
+  height: 100%;
+  padding: 2rem;
+  padding-top: 3rem;
+  overflow: auto;
+  margin-bottom: 3rem !important;
+
+  .minw-0 {
+    min-width: 0;
+  }
+
+  h3 {
+    font-size: 1rem;
+    font-weight: $jupiter-font-weight;
+    color: $p-800;
+
+    * {
+      color: inherit;
+    }
+  }
+
+  p {
+    line-height: inherit;
+  }
+}
