+++ packages/manager/fragments/account-sidebar/tsconfig.json
@@ -22,7 +22,8 @@
     "src/**/*.tsx",
     "src/**/*.vue",
     "tests/**/*.ts",
-    "tests/**/*.tsx"
+    "tests/**/*.tsx",
+    "src/components/AccountSidebar.jsx"
   ],
   "exclude": ["node_modules"]
 }
